+++ CHANGES.rst
@@ -6,6 +6,7 @@ Development
 * Fixed re-publishing issues when certain options change (e.g. parent page)
 * Improve support when using the sphinxcontrib-video extension
 * Introduce the ``confluence_permit_raw_html`` option
+* Provide quirk for CDATA issues on Confluence 8.0.x to 8.2.x
 * Support configuring the theme on generated code block macros
 * Support page-specific editor and full-width overrides
 * Support page-specific parent identifier overrides when publishing